requirements.txt
absl-py==0.9.0
alabaster==0.7.12
argh==0.26.2
astor==0.8.1
atomicwrites==1.4.0
Babel==2.8.0
backcall==0.2.0
bleach==3.1.5
blinker==1.4
brotlipy==0.7.0
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
colorama==0.4.3
decorator==4.4.2
defusedxml==0.6.0
docutils==0.16
entrypoints==0.3
flake8==3.8.3
Flask==1.1.2
future==0.18.2
gast==0.2.2
google-auth-oauthlib==0.4.1
google-pasta==0.2.0
h5py==2.10.0
imagesize==1.2.0
itsdangerous==1.1.0
Jinja2==2.11.2
Keras==2.3.1
Keras-Preprocessing==1.1.0
lazy-object-proxy==1.4.3
mccabe==0.6.1
mkl-fft==1.1.0
mkl-random==1.1.1
mkl-service==2.3.0
nbformat==5.0.7
oauthlib==3.1.0
opt-einsum==3.1.0
packaging==20.4
pandas==1.1.1
pandocfilters==1.4.2
paramiko==2.7.1
parso==0.7.0
pathtools==0.1.2
Pillow==7.2.0
pluggy==0.13.1
protobuf==3.13.0
pyasn1==0.4.8
pyasn1-modules==0.2.7
pycodestyle==2.6.0
pyflakes==2.2.0
Pygments==2.6.1
PyJWT==1.7.1
pyparsing==2.4.7
pyreadline==2.1
pyrsistent==0.16.0
python-dateutil==2.8.1
pytz==2020.1
pywin32==227
PyYAML==5.3.1
pyzmq==19.0.1
QDarkStyle==2.8.1
QtAwesome==0.7.2
QtPy==1.9.0
requests-oauthlib==1.3.0
rope==0.17.0
Rtree==0.9.4
six==1.15.0
snowballstemmer==2.0.0
sortedcontainers==2.2.2
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==1.0.3
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.4
tensorboard==2.2.1
tensorboard-plugin-wit==1.6.0
tensorflow==2.1.0
tensorflow-estimator==2.1.0
termcolor==1.1.0
testpath==0.4.4
tornado==6.0.4
traitlets==4.3.3
typed-ast==1.4.1
ujson==1.35
webencodings==0.5.1
Werkzeug==1.0.1
win-inet-pton==1.1.0
wincertstore==0.2
wrapt==1.11.2
zipp==3.1.0
